requirements.txt
fastapi
pyyaml
pytest
httpx
uvicorn